Staking Change
This is a staking change as proposed by Sols. This new system would work off a 24 coin hour period and no longer rely on the network weight to generate staking rewards. This would be to help the smaller coin holders feel more rewarded and allow them to get their staking rewards faster. (Work in progress)